power: msm_battery: Report charging status reported by the MODEM
MODEM indicates the status of charging via an additional bit
in the RPC reply. Update the charging status based on this bit.
This includes the case when EOC is reported and the charger
is still connected.

(cherry picked from commit 991fb85ccfb627b012b7da8be47b76238fcec9b0)

CRs-Fixed: 429512
Change-Id: I472df4eeaf510a1cc8e62bcf9f0fc57ad61aa1fe
Signed-off-by: Anirudh Ghayal <aghayal@codeaurora.org>