msm: audio  VOIP fixes.
-msm: qdsp5: Moving initialization of wakelock and mutex to mvs init

Added change to move initialization of wakelock and mutex to mvs init
to reduce the VoIP startup time.

-msm: qdsp5: Exit VOIP thread at the end of MVS session

In the MVS open, a thread is created but the same is not getting exited
during the MVS release
Added logic to exit the thread in the release.

-msm: qdsp5: Fix for noise issue during VOIP call

Driver discards the downlink packets when input queue is full.
Added code to block processing of downlink packets in the write()
till a input queue free slot is available for write.

-msm: qdsp5v2: Fix for noise issue during VOIP call

Driver discards the downlink packets when input queue is full.
Added code to block processing of downlink packets in the write()
till a input queue free slot is available for write.

Change-Id: I8f238b572e5ee9f4321089f9fa87dd917098df0d
Signed-off-by: Kalyani Polepeddy  <kpolep@codeaurora.org>