misc: pmem: Explictly mark DMA allocations as GFP_KERNEL
As already applied on ion_cma_heap:
https://www.codeaurora.org/cgit/quic/la/kernel/msm/commit/drivers/gpu/ion?h=kk_3.5&id=83062cf0351637afc7c20456a43ac9b12ffab28b

For legacy reasons, the GFP allocation for dma allocations is currently
marked as 0, which corresponds to atomic allocations. These allocations
should not be marked as atomic since the atomic pool is very small.
Switch to properly using GFP_KERNEL to always allocate from the
correct pool.